requirements.txt
fastapi
uvicorn
python-dotenv
requests
phonenumbers
prisma
langchain-openai
langgraph